+++ docs/changelog.rst
@@ -20,6 +20,9 @@ v0.2 (not yet released)
 * Added iterators to all truth tables: ``begin``, ``end``, ``rbegin``, ``rend``, ``cbegin``, ``cend``, ``crbegin``, ``crend``
   `#14 <https://github.com/msoeken/kitty/pull/14>`_
 
+* Find bits: ``find_first_one_bit``, ``find_last_one_bit``, ``find_first_bit_difference``, ``find_last_bit_difference``
+  `#17 <https://github.com/msoeken/kitty/pull/17>`_
+
 v0.1 (October 22, 2017)
 -----------------------